Delete the old api models and start the work on fixing the resulting errors
WARNING this does NOT RUN and is a work in progress commit.
It still contains the MarketplaceApplication MarketplaceApplicationReference InjectionOption InjectionOptions and TopologyTemplates.
This is most visible in the deployment-completion.service.ts and repository-management.service.ts (should be fixed first).
But also in all files which use those services
