+++ src/app/application-management/application-detail/application-detail-resolver.service.ts
@@ -14,28 +14,18 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
-import { ApplicationDetail } from '../../core/model/application-detail.model';
-import { PlanOperationMetaData } from '../../core/model/planOperationMetaData.model';
 import { ApplicationManagementService } from '../../core/service/application-management.service';
 import { OpenToscaLoggerService } from '../../core/service/open-tosca-logger.service';
-import { Csar } from '../../core/model/new-api/csar.model';
-import { Plan } from '../../core/model/new-api/plan.model';
+import { Csar } from '../../core/model/csar.model';
+import { Plan } from '../../core/model/plan.model';
 
 @Injectable()
 export class ApplicationDetailResolverService {
 
-    resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<ApplicationDetail> {
-        return Observable.forkJoin(
-            [
-                this.appService.getCsarDescriptionByCsarID(route.params['id']).retry(3),
-                this.appService.getBuildPlan(route.params['id']).retry(3),
-                this.appService.getTerminationPlan(route.params['id']).retry(3)
-            ]
-        )
-            .map((result: Array<any>) => new ApplicationDetail(<Csar>result[0], <Plan>result[1], <Plan>result[2]))
-            .catch(reason => {
-                return this.logger.handleError('[application-details-resolver.service][resolve]', reason);
-            });
+    resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<Csar> {
+        return this.appService.getCsarDescriptionByCsarID(route.params['id']).catch(reason => {
+            return this.logger.handleError('[application-details-resolver.service][resolve]', reason);
+        });
     }
 
     constructor(private appService: ApplicationManagementService, private logger: OpenToscaLoggerService) {
