+++ src/app/application-management/application-instance-detail/application-instance-detail.component.html
@@ -4,7 +4,6 @@
             <h1>Instance Details</h1>
             <pre>
                 <code class="language-json">
-                    {{instance | json}}
                 </code>
             </pre>
         </div>
