--- src/app/core/model/application-instance-properties.model.ts
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) 2017 University of Stuttgart.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * and the Apache License 2.0 which both accompany this distribution,
- * and are available at http://www.eclipse.org/legal/epl-v10.html
- * and http://www.apache.org/licenses/LICENSE-2.0
- *
- * Contributors:
- *     Michael Falkenthal - initial implementation
- *     Michael Wurster - initial implementation
- */
-
-import { ResourceReference } from './resource-reference.model';
-
-export class ApplicationInstanceProperties {
-    public instanceReference: ResourceReference;
-    public properties: Array<any>;
-    public selfServiceApplicationURL?: string;
-
-    constructor(instancesReference: ResourceReference, properties: Array<any>, selfServiceApplicationURL?: string) {
-        this.instanceReference = instancesReference;
-        this.properties = properties;
-        this.selfServiceApplicationURL = selfServiceApplicationURL;
-    }
-}
