+++ src/app/core/model/csar-list-entry.model.ts
