+++ src/app/core/model/csar-upload-request.model.ts
