+++ src/app/core/model/csar.model.ts
