+++ src/app/core/model/interface.model.ts
