+++ src/app/core/model/link-reference.ts
