--- src/app/core/model/new-api/plan-instance.model.ts
@@ -1,22 +1,0 @@
-/**
- * Copyright (c) 2017 University of Stuttgart.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * and the Apache License 2.0 which both accompany this distribution,
- * and are available at http://www.eclipse.org/legal/epl-v10.html
- * and http://www.apache.org/licenses/LICENSE-2.0
- *
- * Contributors:
- *     Michael Falkenthal - initial implementation
- */
-
-import { ResourceSupport } from './resource-support.model';
-import { PlanParameter } from '../plan-parameter.model';
-import { PlanLogEntry } from './plan-log-entry.model';
-
-export class PlanInstance extends ResourceSupport {
-    id: string;
-    state: string;
-    output: Array<PlanParameter>;
-    logs: Array<PlanLogEntry>;
-}
