+++ src/app/core/model/operation.model.ts
