+++ src/app/core/model/plan-log-entry.model.ts
