+++ src/app/core/model/resource-support.model.ts
