+++ src/app/core/model/service-template-instance-list-entry.model.ts
