+++ src/app/core/model/service-template-instance-list.model.ts
