+++ src/app/core/model/service-template-instance.model.ts
