+++ src/app/core/model/service-template-list-entry.model.ts
