+++ src/app/core/model/service-template-list.model.ts
