+++ src/app/core/model/service-template.model.ts
