--- src/app/core/model/topology-documentation-element.model.ts
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) 2017 University of Stuttgart.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * and the Apache License 2.0 which both accompany this distribution,
- * and are available at http://www.eclipse.org/legal/epl-v10.html
- * and http://www.apache.org/licenses/LICENSE-2.0
- *
- * Contributors:
- *     Karoline Saatkamp - initial implementation
- */
-export class TopologyDocumentationElement {
-
-    content: Array<string>;
-    source: any;
-    lang: string;
-
-}
-
