+++ src/app/core/service/deployment-completion.service.ts
@@ -15,10 +15,6 @@ import {Http, Headers} from '@angular/http';
 import 'rxjs/add/operator/toPromise';
 import * as _ from 'lodash';
 import {OpenToscaLoggerService} from './open-tosca-logger.service';
-import {MarketplaceApplication} from '../model/marketplace-application.model';
-import {InjectionOptions} from '../model/injection-options.model';
-import {InjectionOption} from '../model/injection-option.model';
-import {TopologyTemplate} from '../model/topology-template.model';
 
 @Injectable()
 export class DeploymentCompletionService {
